requirements.txt
line-bot-sdk==1.16.0
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
SQLAlchemy==1.3.1
Werkzeug==0.15.3
gunicorn==19.9.0
itsdangerous==1.1.0